voice-assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0

# AI Services
groq==0.4.1
deepgram-sdk==3.2.7
elevenlabs==0.2.26

# Audio processing
pyaudio==0.2.11
wave==0.0.2
numpy==1.24.3
scipy==1.11.4

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.7

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-json-logger==2.0.7
